Relocate array utils to helpers.collection (#7498)
